Support retrieval of standard claims from a JWT
To support client assertion authentication we need to retrieve the
client ID from the claims in a signed token in order to establish
the client's identity and thus its key or secret with which to
validate the token.

This commit adds a 'decodeClaims' function to return the standard
set of registered claims defined in the JWT spec.