Introduce Payload type and change use of Jwt type
Payload replaces the use of a plain ByteString in the high-level encoding
functions, to facilitate the creation of nested JWTs.

The Jwt type is now a newtype representing an encoded JWT, which breaks
the previous usage, where it was the return type from a decode operation.

The Payload can be either plain bytes or a nested JWT, making use of the
new Jwt type. For a nested type, the 'cty' header value will be set to
'JWT' as required by the spec.