Add support for unsecured JWTs
This change introduces an 'Unsecured' option for the decoded content
type, so that it is obvious when decoding that the content wasn’t
signed or encrypted. Users will get a pattern match warning if they
don’t take this into account.

Similarly parsing the header no longer has to be either a Jwe or a Jwe,
it can also be unsecured.

Having 'none' as a JWS algorithm is confusing, since "signed with no
algorithm" is really unsigned. We knew for sure that something of type
Jws had a signature, it would make dealing with algorithm preferences in
OpenID simpler, for one thing. However, the spec does it this way so
we’re stuck with it.