Fix JWK matching
The keys used to encode and decode may be different so the code should
reflect this. For example, both a public and private RSA key can be used
for encryption but only the private key can be used for decryption.

The code has been rewritten to provide a set of predicate functions
canDecodeJws, canDecodeJwe and so on, which can be used to filter a
list of JWKs to locate a suitable key or keys.