Add JwtEncoding data type
This contains the algorithm information needed to encode a JWT as either
a JWS or a JWE. It can also optionally be used for validation during
decoding when the required algorithms are known in advance (as is usually
the case).

Decoding an unsecured JWT now explicitly requires that this encoding is
requested or an error will be returned. It is still necessary to pattern
match on the 'Unsecured' JWT content type, but this provides an additional
safeguard against header substitution attacks, since an end user can't
just pattern match and pull out the content without thinking about whether
the content is signed or not.