Internal crypto improvements
The AES functions will throw exceptions if the key, IV is the wrong
length for the algorithm being used. This change generates a dummy key
and IV and uses these instead if decrypting the CEK fails or the lengths
are wrong. It also checks the payload length for CBC mode to make sure
it is a multiple of the block size, since this can also cause an
exception. This should also make the timing more uniform, since a
payload decryption will always be performed, even if the content key
fails to decrypt.