Remove dependency on aeson-qq
Replace inline JSON with a file.

This package pulls in lots of other dependencies which are slow to build
and fragile (haskell-src-exts, template haskell stuff).

The project won't build in stackage nightly just now because of these.

Since it's only used in one place, the simplest option is to remove it.