Check key lengths when selecting keywrap keys
When selecting a JWK from a list to be used with AES keywrap, we should check
the length of the key matches the algorithm.

Also fix JWE tests by providing valid keys for AES192 and AES256 (the JWK from
Appendix 3 is for AES128).

This came to light because of changes in cryptonite 0.21 which includes stricter
key length checking. Building against 0.21 in stackage caused the tests to fail.

See:

https://github.com/haskell-crypto/cryptonite/issues/113
https://github.com/fpco/stackage/issues/2162