Implement a parser for JWTs as the first stage in decoding
Currently the decryption and validity checking of the various parts of the
JWT are somwehat combined, which makes things a bit messy.

The aim is to run the JWT through the parser first and remove as many error
cases as possible at this stage before going on to decode the data. Future
changes can probably make better use of the type system and extensions to do a
better job of this.

For now this is only used in decoding JWEs. It creates specific data types for
the IV and auth tag and checks that they are consistent with the enc algorithm.
It may also be possible to detect whether the encrypted CEK is valid, based on
the length of the data and the key encryption algorithm used (for RSA it would
also depend on the key size).

The parser just returns BadCrypto when it fails for consistency with the current
errors. A future major version should probably introduce a ParseError.