Export EcCurve type from Jwk module
This allows construction of Jwk's from scratch (see #24).

The Jwk module was really just intended as a way of decoding JWKs
from JSON and could probably be better designed, but it's also
useful to be able to generate keys or read them from other sources
and encode them as JWKs and this simple change makes this possible
for elliptic curve keys.