Fix broken tests due to change in aeson field order
See https://github.com/commercialhaskell/stackage/issues/5878

The problem is that the JWE header is generated internally and since
it is included in the signature, the tests which attempt to reproduce
the JWE values from the spec exactly fail.

Since the header is a simple structure, instead of using Aeson
to encode the encode it we just create a ByteString directly with
the original order (and encode the Alg and Enc values into it).