Handle unsupported alg and kty values when loading JWKs
The alg and kty values in a JWK may include values which we don't
support - either through choice (such as various EC algorithms) or
because they are currently unknown (the JWA spec allows new values
to be registered for both kty and alg in future).

In cases like these, we don't want loading a key set to fail (as
reported in #31). Instead, if either these fields is preset but
fails to parse to a known value, we return an "UnsupportedJwk"
which just wraps the JSON object which contains the key data.