Get rid of some incomplete pattern warnings
Added compiler options to Jwa.hs and test files. This is OK since
the incomplete matches in Jwa are for name lookups by the Alg and Enc
type. All values of these types should be catered for so we shouldn't
get an error unless a new algorithm is added and we forget to add it to
the names array (a bug). So we should never have an issue at runtime.

Clarified the code in Jwe by adding an explicit error when payload
encryption fails. Again this should never happen as the IV and key
are generated by us and should be valid for the given algorithm.